+++ types/resources/VirtualBankAccount.d.ts
@@ -1,4 +1,5 @@
 ///<reference path='./../core.d.ts'/>
+///<reference path='./../index.d.ts'/>
 declare module 'chargebee' {
   export interface VirtualBankAccount {